refactor(turbopack): let ExcutionEnvironment to be extendable trait
